+++ genvalidity/CHANGELOG.md
@@ -1,2 +1,3 @@
 Unreleased changes:
 * -0 is now a valid value for Double and Float.
+* `genUnchecked :: Gen Double` now also generates invalid values.